ajout du lock dans la classe Emulation + erreur Error remplacé par FatalError dans tout le projet

+++ messages/fight/GameFightOptionToggleMessage.java
@@ -14,7 +14,7 @@ public class GameFightOptionToggleMessage extends Message {
 		this.option = option;
 		
 		ByteArray buffer = new ByteArray();
-		buffer.writeByte((byte) this.option);
+		buffer.writeByte(this.option);
 		completeInfos(buffer);
 	}
 }
